Reference-information-verification/requirements.txt
PyQt5>=5.15.0
requests>=2.28.0
numpy>=1.21.0
faiss-cpu>=1.7.0
python-docx>=0.8.10
PyPDF2>=3.0.0
spacy>=3.5.0